refactor(perf): user message renderer
