+++ package.json
@@ -252,7 +252,7 @@
     "@lobehub/chat-plugin-sdk": "^1.32.4",
     "@lobehub/chat-plugins-gateway": "^1.9.0",
     "@lobehub/desktop-ipc-typings": "workspace:*",
-    "@lobehub/editor": "^4.0.0",
+    "@lobehub/editor": "^4.2.0",
     "@lobehub/icons": "^5.0.0",
     "@lobehub/market-sdk": "^0.31.3",
     "@lobehub/tts": "^5.1.2",
